Implement the voltage calibration in the uproot backend as its own class.
